app/algorithm/pyproject.toml
[project]
name = "algorithm"
version = "0.1.0"
description = "Сервис построения графика архитектуры приложения"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.7",
    "ast2json>=0.4",
    "dotenv>=0.9.9",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.7",
    "networkx>=3.5",
    "pandas>=2.3.3",
    "protobuf>=6.33.0",
    "pyvis>=0.3.2",
    "scipy>=1.16.2",
]
